tipologia crimini Nicolò
tipologia crimini service serviceimpl control restcontrol

+++ ReportTracker/src/main/java/com/repositories/TipologiaCrimineRepository.java
@@ -1,11 +1,12 @@
 package com.repositories;
+
+
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import org.springframework.stereotype.Repository;
-
 import com.model.TipologiaCrimine;
 
 @Repository 
-public interface TipologiaCrimineRepository extends JpaRepository <TipologiaCrimine,Long> {
+public interface TipologiaCrimineRepository extends JpaRepository <TipologiaCrimine,Integer> {
 
 }
+
